Rename ccopen.py to ccio.py to avoid name clashes
