+++ src/cclib/parser/__init__.py
@@ -34,7 +34,7 @@ from .qchemparser import QChem
 # This allow users to type:
 #   from cclib.parser import ccopen
 #   from cclib.parser import ccread
-from .ccopen import ccopen
-from .ccopen import ccread
+from .ccio import ccopen
+from .ccio import ccread
 
 from .data import ccData
