+++ src/cclib/parser/utils.py
@@ -10,7 +10,7 @@
 # received a copy of the license along with cclib. You can also access
 # the full license online at http://www.gnu.org/copyleft/lgpl.html.
 
-"""Utilities often used by cclib parsers"""
+"""Utilities often used by cclib parsers and scripts"""
 
 
 def convertor(value, fromunits, tounits):
